\name{RcppClassic-package}
\alias{RcppClassic-package}
\alias{RcppClassic}
\docType{package}
\title{Deprecated R / C++ interface}
\description{The \pkg{RcppClassic} package provides the classic Rcpp API. }
\author{Dominick Samperi wrote the initial versions of Rcpp (and
  RcppTemplate) during 2005 and 2006.  Dirk Eddelbuettel made some
  additions, and became maintainer in 2008. Dirk Eddelbuettel and Romain
  Francois have been extending Rcpp since 2009.
}
\keyword{programming}
\keyword{interface}
